requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
cssutils==2.11.1
Flask==3.1.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
more-itertools==10.7.0
packaging==25.0
premailer==3.10.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
yagmail==0.15.293
pip install gunicorn
pip freeze > requirements.txt